Remove the transport sub-module
The `transport` module is tiny, we can just put that code in `mod.rs` with no
loss of clarity.
